[Refactor] cylinder triangulation
Thanks to some line-profiling it was clear that the most time-consuming
part of rendering a set of cylinders was a python for-loop over each
cylinder.
Since the cylinder vertices were already stored in scipp variables, it
was straightforward to modify the triangulation routine to act on them
directly.

Triangulation now occurs for all cylinders simultaneously, which
produced a 90x speed-up for the profiled test case of 13500 cylinders.

Large numbers of cylinders (10^6, at least) with default settings can
cause an error in the communication layer of jupyter. It's not clear at
this point if that is a result of a software configuration or hardware
limitations on my machine.
This limit can be overcome by reducing the number of vertices per
cylinder.